First render of something

+++ src/camera.cuh
@@ -4,7 +4,7 @@
 #include "vector.cuh"
 
 struct Camera {
-    Camera(vec3 position, vec3 direction, vec2 resolution, float focal_length) {
+    Camera(vec3 position, vec3 direction, ivec2 resolution, float focal_length) {
         this->position      = position;
         this->direction     = direction;
         this->resolution    = resolution;
@@ -12,7 +12,7 @@ struct Camera {
     }
     vec3 position;
     vec3 direction;
-    vec2 resolution;
+    ivec2 resolution;
     float focal_length;
 };
